+++ spec/telegram/bot/updates_controller/translation_spec.rb
@@ -14,4 +14,18 @@ RSpec.describe Telegram::Bot::UpdatesController::Translation do
       controller.t('.hello')
     end
   end
+
+  describe described_class::ClassMethods do
+    describe '#translate' do
+      let(:controller_class) do
+        double(controller_path: 'telegram/webhooks').
+          tap { |x| x.extend(described_class) }
+      end
+
+      it 'uses controller_path for lazy translations' do
+        expect(I18n).to receive(:translate).with('telegram.webhooks.hello', {})
+        controller_class.t('.hello')
+      end
+    end
+  end
 end